Moves KML ops to the plugin gama.extension.kml
