+++ gama.api/src/gama/api/types/file/IGamaFile.java
@@ -29,20 +29,20 @@ import gama.api.utils.geometry.IEnvelopeProvider;
 
 /**
  * The root interface for all file types in the GAMA modeling platform.
- * 
+ *
  * <p>
- * {@code IGamaFile} provides a unified abstraction for accessing and manipulating files within GAMA models.
- * It extends both {@link gama.api.types.misc.IContainer.Addressable} and
- * {@link gama.api.types.misc.IContainer.Modifiable} to enable file contents to be treated as addressable
- * and modifiable containers, allowing natural integration with GAML's container operations.
+ * {@code IGamaFile} provides a unified abstraction for accessing and manipulating files within GAMA models. It extends
+ * both {@link gama.api.types.misc.IContainer.Addressable} and {@link gama.api.types.misc.IContainer.Modifiable} to
+ * enable file contents to be treated as addressable and modifiable containers, allowing natural integration with GAML's
+ * container operations.
  * </p>
- * 
+ *
  * <h2>Type Parameters</h2>
  * <ul>
  * <li>{@code C} - The container type used to store file contents (e.g., IList, IMatrix, IMap)</li>
  * <li>{@code Contents} - The type of individual elements within the container</li>
  * </ul>
- * 
+ *
  * <h2>Key Responsibilities</h2>
  * <ul>
  * <li><b>File Metadata:</b> Provides access to file properties (name, path, extension, existence)</li>
@@ -51,7 +51,7 @@ import gama.api.utils.geometry.IEnvelopeProvider;
  * <li><b>I/O Operations:</b> Supports both reading and writing file data</li>
  * <li><b>Attributes:</b> Provides access to file-specific metadata (e.g., CSV headers, shapefile fields)</li>
  * </ul>
- * 
+ *
  * <h2>File Attributes</h2>
  * <p>
  * All files expose the following attributes accessible in GAML via dot notation:
@@ -67,7 +67,7 @@ import gama.api.utils.geometry.IEnvelopeProvider;
  * <li>{@code attributes} - File-specific attributes (varies by file type)</li>
  * <li>{@code contents} - The file contents as a container</li>
  * </ul>
- * 
+ *
  * <h2>Lifecycle</h2>
  * <ol>
  * <li><b>Creation:</b> File object is created with a path (local or URL)</li>
@@ -76,7 +76,7 @@ import gama.api.utils.geometry.IEnvelopeProvider;
  * <li><b>Buffering:</b> Once loaded, contents remain in memory (buffer)</li>
  * <li><b>Writing:</b> Modified contents can be flushed back to disk</li>
  * </ol>
- * 
+ *
  * <h2>Specialized File Types</h2>
  * <p>
  * GAMA provides specialized interfaces and implementations for specific file types:
@@ -85,35 +85,36 @@ import gama.api.utils.geometry.IEnvelopeProvider;
  * <li>{@link Drawable} - Marker interface for files that can be displayed visually</li>
  * <li>{@link WithGeometry} - Files containing geospatial data (e.g., shapefiles, GeoJSON)</li>
  * </ul>
- * 
+ *
  * <h2>Usage Example</h2>
+ *
  * <pre>{@code
  * // In Java - working with a file
  * IGamaFile<IList<String>, String> file = new GenericFile(scope, "data.txt");
- * 
+ *
  * // Check if file exists
  * if (file.exists(scope)) {
  *     // Get contents as a list
  *     IList<String> lines = file.getContents(scope);
- *     
+ *
  *     // Iterate over lines
  *     for (String line : lines) {
  *         System.out.println(line);
  *     }
- *     
+ *
  *     // Modify and save
  *     lines.add("New line");
  *     file.setContents(lines);
  *     // Flush to disk via save operation
  * }
- * 
+ *
  * // In GAML - natural container syntax
  * file my_file <- file("data.txt");
  * loop line over: my_file {
  *     write line;
  * }
  * }</pre>
- * 
+ *
  * <h2>Implementation Notes</h2>
  * <ul>
  * <li>Implementations should handle both local file paths and URLs</li>
@@ -121,17 +122,17 @@ import gama.api.utils.geometry.IEnvelopeProvider;
  * <li>Buffering should be lazy to avoid unnecessary I/O</li>
  * <li>Error handling should provide clear feedback about file issues</li>
  * </ul>
- * 
+ *
  * @param <C>
  *            the container type for storing file contents (must be both Addressable and Modifiable)
  * @param <Contents>
  *            the type of individual elements within the container
- * 
+ *
  * @see gama.api.types.file.GamaFile
  * @see gama.api.types.file.GenericFile
  * @see gama.api.types.misc.IContainer
  * @see gama.api.gaml.types.GamaFileType
- * 
+ *
  * @author Alexis Drogoul
  * @since GAMA 1.0
  */
@@ -184,13 +185,12 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 	/**
 	 * Marker interface for file types that can be rendered visually in displays.
-	 * 
+	 *
 	 * <p>
-	 * Files implementing this interface can be directly used in GAML display statements
-	 * (e.g., image files, 3D mesh files). This allows them to be drawn or rendered in
-	 * graphical displays without additional conversion.
+	 * Files implementing this interface can be directly used in GAML display statements (e.g., image files, 3D mesh
+	 * files). This allows them to be drawn or rendered in graphical displays without additional conversion.
 	 * </p>
-	 * 
+	 *
 	 * @see gama.api.ui.displays.IAsset
 	 */
 	interface Drawable {
@@ -199,24 +199,23 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 	/**
 	 * Specialized interface for files containing geospatial/geometric data.
-	 * 
+	 *
 	 * <p>
-	 * Files implementing this interface contain shape/geometry information and can provide
-	 * their contents as a list of {@link IShape} objects. This includes file types like:
+	 * Files implementing this interface contain shape/geometry information and can provide their contents as a list of
+	 * {@link IShape} objects. This includes file types like:
 	 * </p>
 	 * <ul>
 	 * <li>Shapefiles (.shp)</li>
 	 * <li>GeoJSON files (.geojson, .json)</li>
-	 * <li>KML/KMZ files</li>
 	 * <li>OSM (OpenStreetMap) files</li>
 	 * <li>GeoTIFF files (with vector data)</li>
 	 * </ul>
-	 * 
+	 *
 	 * <p>
-	 * These files are both {@link Drawable} (can be displayed) and provide geometric
-	 * operations through the {@link IShape} interface.
+	 * These files are both {@link Drawable} (can be displayed) and provide geometric operations through the
+	 * {@link IShape} interface.
 	 * </p>
-	 * 
+	 *
 	 * @see gama.api.types.geometry.IShape
 	 * @see gama.api.utils.geometry.IEnvelopeProvider
 	 */
@@ -224,12 +223,11 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 		/**
 		 * Returns the combined geometry of all shapes in this file.
-		 * 
+		 *
 		 * <p>
-		 * This method computes and returns a single {@link IShape} that represents
-		 * the union or collection of all geometric features contained in the file.
-		 * For files with multiple features, this is typically a multi-geometry or
-		 * geometry collection.
+		 * This method computes and returns a single {@link IShape} that represents the union or collection of all
+		 * geometric features contained in the file. For files with multiple features, this is typically a
+		 * multi-geometry or geometry collection.
 		 * </p>
 		 *
 		 * @param scope
@@ -248,17 +246,16 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 	/**
 	 * Sets whether this file should be treated as writable.
-	 * 
+	 *
 	 * <p>
-	 * This method controls whether the file can accept modifications and be saved back to disk.
-	 * Setting a file as writable enables operations like {@link #setContents(IContainer)} and
-	 * allows the file to be used as an output destination in save operations.
+	 * This method controls whether the file can accept modifications and be saved back to disk. Setting a file as
+	 * writable enables operations like {@link #setContents(IContainer)} and allows the file to be used as an output
+	 * destination in save operations.
 	 * </p>
-	 * 
+	 *
 	 * <p>
-	 * Note: This affects the file object's behavior, not the actual file system permissions.
-	 * A file that is writable in the object sense may still fail to write if the underlying
-	 * file system permissions don't allow it.
+	 * Note: This affects the file object's behavior, not the actual file system permissions. A file that is writable in
+	 * the object sense may still fail to write if the underlying file system permissions don't allow it.
 	 * </p>
 	 *
 	 * @param scope
@@ -270,16 +267,16 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 	/**
 	 * Sets the contents of this file.
-	 * 
+	 *
 	 * <p>
-	 * Replaces the current buffered contents with the provided container. This operation
-	 * updates the in-memory representation but does not immediately write to disk. To persist
-	 * changes, the file must be flushed using appropriate save operations.
+	 * Replaces the current buffered contents with the provided container. This operation updates the in-memory
+	 * representation but does not immediately write to disk. To persist changes, the file must be flushed using
+	 * appropriate save operations.
 	 * </p>
-	 * 
+	 *
 	 * <p>
-	 * The file must be writable for this operation to succeed. The provided container should
-	 * match the expected type for this file (e.g., IList for text files, IMatrix for grid files).
+	 * The file must be writable for this operation to succeed. The provided container should match the expected type
+	 * for this file (e.g., IList for text files, IMatrix for grid files).
 	 * </p>
 	 *
 	 * @param cont
@@ -291,10 +288,10 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 	/**
 	 * Creates a copy of this file.
-	 * 
+	 *
 	 * <p>
-	 * Returns a new file instance with the same path and properties as this file.
-	 * The buffer contents are shared or copied depending on the implementation.
+	 * Returns a new file instance with the same path and properties as this file. The buffer contents are shared or
+	 * copied depending on the implementation.
 	 * </p>
 	 *
 	 * @param scope
@@ -306,13 +303,12 @@ public interface IGamaFile<C extends IContainer.Modifiable, Contents>
 
 	/**
 	 * Gets the internal buffer containing the file's loaded contents.
-	 * 
+	 *
 	 * <p>
-	 * Returns the cached container holding the file contents. If the file hasn't been
-	 * loaded yet, this may return null. Use {@link #getContents(IScope)} to ensure
-	 * the file is loaded before accessing contents.
+	 * Returns the cached container holding the file contents. If the file hasn't been loaded yet, this may return null.
+	 * Use {@link #getContents(IScope)} to ensure the file is loaded before accessing contents.
 	 * </p>
-	 * 
+	 *
 	 * <p>
 	 * Direct buffer access should be used cautiously as it bypasses lazy loading mechanisms.
 	 * </p>
