+++ gama.core/plugin.xml
@@ -64,10 +64,6 @@
             description="Saves graphs in different formats (from JGraphT)">
       </implementation>
       <implementation
-            class="gama.gaml.statements.save.KmlSaver"
-            description="Saves .kml or .kmz files">
-      </implementation>
-      <implementation
             class="gama.gaml.statements.save.ShapeSaver"
             description="Saves geometries and agents as shapefiles">
       </implementation>
