+++ gama.dependencies/.classpath
@@ -19,7 +19,6 @@
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jai_imageio-1.1.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/jai_core 1.1.3/jai_core-1.1.3.jar" sourcepath="libs/jai_core 1.1.3/jai_core-1.1.3-sources .zip"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jai_codec-1.1.3.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/jak 3.0.11/javaapiforkml-3.0.11.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/kabeja 0.5/kabeja-0.5.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/xml-commons-resolver-1.2.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/systems-common-2.1.jar"/>
