+++ gama.dependencies/META-INF/MANIFEST.MF
@@ -232,12 +232,6 @@ Export-Package: COM.ibm.db2.app,
  com.sun.media.jai.tilecodec,
  com.sun.media.jai.util,
  com.sun.media.jai.widget,
- de.micromata.opengis.kml.v_2_2_0,
- de.micromata.opengis.kml.v_2_2_0.annotations,
- de.micromata.opengis.kml.v_2_2_0.atom,
- de.micromata.opengis.kml.v_2_2_0.gx,
- de.micromata.opengis.kml.v_2_2_0.util,
- de.micromata.opengis.kml.v_2_2_0.xal,
  dk.abj.svg.action,
  edu.umd.cs.findbugs.annotations,
  gama.dependencies,
@@ -1750,7 +1744,6 @@ Bundle-ClassPath: .,
  libs/geotools 33.4/systems-common-2.1.jar,
  libs/geotools 33.4/xml-commons-resolver-1.2.jar,
  libs/kabeja 0.5/kabeja-0.5.0.jar,
- libs/jak 3.0.11/javaapiforkml-3.0.11.jar,
  libs/jai-imageio-core 1.4.0/jai-imageio-core-1.4.0.jar,
  libs/geotools 33.4/jai_codec-1.1.3.jar,
  libs/jai_core 1.1.3/jai_core-1.1.3.jar,
