+++ gama.dependencies/build.properties
@@ -146,7 +146,6 @@ bin.includes = .,\
                libs/geotools 33.4/systems-common-2.1.jar,\
                libs/geotools 33.4/xml-commons-resolver-1.2.jar,\
                libs/kabeja 0.5/kabeja-0.5.0.jar,\
-               libs/jak 3.0.11/javaapiforkml-3.0.11.jar,\
                libs/jai-imageio-core 1.4.0/jai-imageio-core-1.4.0.jar,\
                libs/geotools 33.4/jai_codec-1.1.3.jar,\
                libs/jai_core 1.1.3/jai_core-1.1.3.jar,\
