+++ gama.extension.kml/.settings/org.eclipse.jdt.apt.core.prefs
@@ -1,0 +1,4 @@
+eclipse.preferences.version=1
+org.eclipse.jdt.apt.aptEnabled=true
+org.eclipse.jdt.apt.genSrcDir=gaml
+org.eclipse.jdt.apt.reconcileEnabled=false
