+++ gama.extension.kml/build.properties
@@ -1,0 +1,10 @@
+source.. = src/,\
+           gaml/
+output.. = bin/
+jre.compilation.profile = JavaSE-25
+bin.includes = .,\
+               plugin.xml,\
+               META-INF/,\
+               models/,\
+               libs/jak 3.0.11/javaapiforkml-3.0.11.jar
+javacDefaultEncoding.. = UTF-8
