+++ gama.extension.kml/src/gama/extension/kml/GamaKmlExport.java
@@ -8,7 +8,7 @@
  * Visit https://github.com/gama-platform/gama for license information and contacts.
  *
  ********************************************************************************************************/
-package gama.core.util.file;
+package gama.extension.kml;
 
 import java.io.File;
 import java.io.IOException;
