+++ gama.feature.core/feature.xml
@@ -32,19 +32,59 @@
       <import plugin="org.eclipse.equinox.p2.metadata"/>
       <import plugin="org.eclipse.equinox.p2.operations"/>
       <import plugin="org.eclipse.equinox.p2.engine"/>
-      <import plugin="gama.extension.traffic"/>
-      <import plugin="gama.extension.database"/>
-      <import plugin="gama.extension.network"/>
-      <import plugin="gama.extension.dataframe"/>
       <import plugin="org.eclipse.xtext"/>
       <import plugin="org.eclipse.xtext.util"/>
       <import plugin="org.eclipse.emf.common"/>
       <import plugin="org.eclipse.xtext.xbase.lib"/>
-      <import plugin="gama.extension.maths"/>
-      <import plugin="gama.extension.batch"/>
-      <import plugin="gama.dependencies"/>
-      <import plugin="gama.workspace"/>
    </requires>
+
+   <plugin
+         id="gama.dependencies"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.api"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.workspace"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.dataframe"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.osm"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.kml"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.database"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.traffic"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.network"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.maths"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.batch"
+         version="0.0.0"/>
+
+   <plugin
+         id="gama.extension.stats"
+         version="0.0.0"/>
 
    <plugin
          id="gama.core"
@@ -67,27 +107,7 @@
          version="0.0.0"/>
 
    <plugin
-         id="gama.extension.stats"
-         version="0.0.0"/>
-
-   <plugin
-         id="gama.extension.batch"
-         version="0.0.0"/>
-
-   <plugin
          id="gama.dev"
-         version="0.0.0"/>
-
-   <plugin
-         id="gama.api"
-         version="0.0.0"/>
-
-   <plugin
-         id="gama.workspace"
-         version="0.0.0"/>
-
-   <plugin
-         id="gama.extension.osm"
          version="0.0.0"/>
 
 </feature>
